LayoutTests/imported/w3c/web-platform-tests/trusted-types/navigate-to-javascript-url-003-expected.txt


FAIL Setting window.location to a javascript: URL with a default policy that throws should report a CSP violation without rethrowing the exception. assert_equals: expected 1 but got 0
